% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfa_dl.R
\name{bfa_dl}
\alias{bfa_dl}
\title{Bayesian Factor Analysis and Covariance Estimation with a row-wise DL prior}
\usage{
bfa_dl(
  X,
  iter_warmup = 1000,
  iter_sampling = 1000L,
  k = NULL,
  prior_params = list(a_sigma = NULL, b_sigma = NULL, a = NULL),
  verbose = TRUE
)
}
\arguments{
\item{X}{Data matrix}

\item{iter_warmup}{Number of warmup iterations to run per chain.}

\item{iter_sampling}{Number of post-warmup iterations to run per chain.}

\item{k}{Number of factors (fixed)}

\item{prior_params}{Parameters of the DL prior.}

\item{verbose}{(logical) Show progress bar?}
}
\value{
List with samples for each parameter
}
\description{
Bayesian Factor Analysis and Covariance Estimation with a row-wise DL prior
}
\examples{
set.seed(219)

# 0. Data simulation as in Bhattacharya & Dunson (2011)
sim <- sim_data_bfa(n = 200, p = 20, k = 3)

# 1. Fixed number of factors (no adaptation in the DL case)
fit_dl_fixed <- bfa_dl(sim$X, k = 6)
plots_Sigma_X <- plot_Sigma_X(fit_dl_fixed, real = sim$Sigma_X_true)
plots_Sigma_X$main
plots_Sigma_X$residual
plots_Sigma_X$scatter

aligned <- match_align(fit_dl_fixed)
plot_match_align(aligned, "Lambda", "mean")
plot_match_align(aligned, "Eta", "mean")
}
