% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfa_mgps.R
\name{bfa_mgps}
\alias{bfa_mgps}
\title{Bayesian Factor Analysis with an MGPS prior}
\usage{
bfa_mgps(
  X,
  iter_warmup = 1000,
  iter_sampling = 1000,
  adapt = TRUE,
  k_init = NULL,
  prior_params = list(a_sigma = NULL, b_sigma = NULL, nu = NULL, a1 = NULL, a2 = NULL,
    alpha0 = NULL, alpha1 = NULL),
  eps = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{X}{Data matrix}

\item{iter_warmup}{Number of warmup iterations to run per chain.}

\item{iter_sampling}{Number of post-warmup iterations to run per chain.}

\item{adapt}{(logical) Adapt number of factors across samples?}

\item{k_init}{If \code{adapt = FALSE}, the persistent number of factors. If
\code{adapt = TRUE}, the initial number of factors in the sampler. If \code{NULL}
the number of factors is selected automatically as \code{3*log(p)} where \code{p}
is the number of columns of \code{X}.}

\item{prior_params}{Parameters of the MGPS prior.}

\item{eps}{If \code{adapt = TRUE}, tolerance for column sums of \code{Lambda} in the
adaptation step.}

\item{verbose}{(logical) Show progress bar?}
}
\value{
List with samples for each parameter
}
\description{
Bayesian Factor Analysis with an MGPS prior
}
\examples{
NULL
}
