% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_align_algorithm.R
\name{match_align}
\alias{match_align}
\title{Resolve rotational ambiguity with the MatchAlign algortithm}
\usage{
match_align(fit, orth_procedure = "varimax")
}
\arguments{
\item{fit}{Returned samples from any \verb{bfa_*()} and \verb{blfr_*()} functions with fixed number of factors.}

\item{orth_procedure}{Orthogonalization procedure to use. Currently only supports VARIMAX.}
}
\value{
List of rotationally aligned factor loadings and factors samples
}
\description{
Performs the varimax rotation on the factor loadings samples and
column-based matching to resolve resultant sign and label switching.
Rotates the factors along with the loadings to induce identifiability jointly.
}
\examples{
set.seed(219)

# 0. Data simulation as in Bhattacharya & Dunson (2011)
sim <- sim_data_bfa(n = 200, p = 20, k = 3)

# 1. Adaptive number of factors, starting at floor(3*log(p)) = 8
fit_cusp <- bfa_cusp(sim$X)
k_fit <- summary_k(fit_cusp, real = sim$k_true) # returns median too

# 2. Fixed number of factors -> MatchAlign
fit_cusp_fixed <- bfa_mgps(sim$X, adapt = FALSE, k_init = k_fit)
aligned <- match_align(fit_cusp_fixed)
plot_match_align(aligned, "Lambda", "mean")
plot_match_align(aligned, "Eta", "mean")
}
