% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Delta_X}
\alias{plot_Delta_X}
\title{Summary Plots of Induced Covariate and Predictor effects.}
\usage{
plot_Delta_X(fit_blfr, real = NULL, X_names = NULL, Z_names = NULL)
}
\arguments{
\item{fit_blfr}{Returned samples from any \verb{blfr_*()} function with covariates}

\item{real}{Real matrix of induced covariate and predictor interaction effects,
if known}

\item{X_names}{Optional character vector of predictor names, of length equal
to the number of variables in \code{X}.}

\item{Z_names}{Optional character vector of covariate names, of length equal
to the number of covariates \code{q}}
}
\value{
Named list of plots
}
\description{
Summary Plots of Induced Covariate and Predictor effects.
}
\examples{
NULL
}
