% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_effects_Z}
\alias{plot_effects_Z}
\title{Summary Plots of Covariate Effects}
\usage{
plot_effects_Z(fit_blfr, real = NULL, Z_names = NULL)
}
\arguments{
\item{fit_blfr}{Returned samples from any \verb{blfr_*()} function with covariates}

\item{real}{Real vector of covariate effects, if known}

\item{Z_names}{Optional character vector of covariate names, of length equal
to the number of covariates \code{q}.}
}
\value{
Named list of plots
}
\description{
Summary Plots of Covariate Effects
}
\examples{
NULL
}
