% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Omega_X}
\alias{plot_Omega_X}
\title{Summary Plots of Induced Interaction Effects}
\usage{
plot_Omega_X(fit_blfr, real = NULL)
}
\arguments{
\item{fit_blfr}{Returned samples from any \verb{blfr_*()} function with interactions}

\item{real}{Real matrix of induced interaction effects, if known}
}
\value{
Named list of plots
}
\description{
Summary Plots of Induced Interaction Effects
}
\examples{
NULL
}
