% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Sigma_X}
\alias{plot_Sigma_X}
\title{Summary Plots of Posterior Covariance}
\usage{
plot_Sigma_X(fit, real = NULL, X_names = NULL, covariance = TRUE)
}
\arguments{
\item{fit}{Returned samples from any \verb{bfa_*()} or \verb{blfr_*()} function.}

\item{real}{Real matrix if known, \code{NULL} otherwise. When \code{covariance = TRUE}
this is the real covariance matrix. When \code{covariance = FALSE} it is the real
correlation matrix.}

\item{X_names}{Optional character vector of covariate names, of length equal
to the number of variables in \code{X}. Used to label axes of plots.}

\item{covariance}{If \code{TRUE} (default), plots the posterior covariance matrix.
If \code{FALSE}, plots the implied correlation matrix instead.}
}
\value{
Named list of plots. If \code{real=NULL}, \code{plots$main} is the plot of the
mean posterior covariance (or correlation) of X. Otherwise, three plots:
\itemize{
\item \code{main}: compares mean posterior estimate with the real matrix
\item \code{residual}: plot of residuals between mean posterior and real matrix
\item \code{scatter}: scatterplot of mean posterior elements vs real elements.
}
}
\description{
Summary Plots of Posterior Covariance
}
\examples{
set.seed(219)

# 0. Data simulation as in Bhattacharya & Dunson (2011)
sim <- sim_data_bfa(n = 200, p = 20, k = 3)

# 1. Adaptive number of factors, starting at floor(3*log(p)) = 8
fit_mgps <- bfa_mgps(sim$X)
plots_Sigma_X <- plot_Sigma_X(fit_mgps, real = sim$Sigma_X_true)
plots_Sigma_X$main
plots_Sigma_X$residual
plots_Sigma_X$scatter
}
