% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{predict_blfr}
\alias{predict_blfr}
\title{Make Predictions with a Latent Regression Model}
\usage{
predict_blfr(fit_blfr, X_test)
}
\arguments{
\item{fit_blfr}{Returned samples from any \verb{blfr_*()} function}

\item{X_test}{\code{n_test} times p matrix of covariates}
}
\value{
\code{n_test} dimensional vector of predictions
}
\description{
Make Predictions with a Latent Regression Model
}
\examples{
NULL
}
