% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data_blfr}
\alias{sim_data_blfr}
\title{Generate Data for Latent Regression Examples}
\usage{
sim_data_blfr(
  n,
  n_test,
  p,
  k,
  interactions = FALSE,
  beta_sd = 2,
  omega_sd = 1,
  sigma2_y = 0.5
)
}
\arguments{
\item{n}{Number of observations}

\item{n_test}{Number of test observations}

\item{p}{Number of covariates in X}

\item{k}{Number of latent factors}

\item{interactions}{(logical) Include interaction effects?}

\item{beta_sd}{standard deviation of latent main effects}

\item{omega_sd}{standard deviation of latent interaction effects}

\item{sigma2_y}{observation level noise in response y}
}
\value{
Named list with the following elements:
\itemize{
\item \code{X}: \eqn{n \times p} matrix of predictors
\item \code{y}: \eqn{n}-dimensional vector of the response
\item \code{X_test}: \eqn{\text{n\_test} \times p} matrix of test predictors
\item \code{y_test}: \code{n_test}-dimensional vector of test responses
\item \code{Lambda}: Simulated factor loadings
\item \code{sigma2}: Vector of intrinsic variances for each covariate
\item \code{beta}: latent main effects
\item \code{Omega}: latent interaction effects, \code{NULL} if \code{interactions=FALSE}
\item \code{sigma2_y}: response level noise
\item \code{Sigma_X_true}: True covariance matrix of \code{X}, i.e.
\eqn{\Lambda^T\lambda + \text{diag}(\sigma^2)}
\item \code{beta_X}: induced main effects in original X scale
\item \code{Omega_X}: induced interaction effects in original X scale
\item \code{intercept_X}: induced intercept, i.e. \eqn{tr(\Omega V)}
\item \code{k_true}: True number of latent factors
}
}
\description{
Generate Data for Latent Regression Examples
}
\details{
Data is generated as follows:
\itemize{
\item Latent level as in Bhattacharya & Dunson (2011): number of non-zero
columns of the matrix of loadings is linearly decreasing from 2k down to
k + 1 and selected at random. Then non-zero elements are sampled
from a N(0, 9) distribution.
\item Latent effects: \code{beta} is drawn from a Normal distribution with mean 0 and
standard deviation \code{beta_sd}. \code{Omega} elements are drawn from a Normal
distribution with mean 0 and sd \code{omega_sd}, then \eqn{\Omega = \frac{\Omega + \Omega^T}{2}}
\item Observations are drawn with mean from the corresponding latent regression
model and variance \code{sigma2_y}.
}
}
\examples{
set.seed(219)

# 0. Data simulation
sim <- sim_data_blfr(n = 200, n_test = 20, p = 20, k = 3)

# 1. Adaptive number of factors, starting at floor(3*log(p)) = 8
fit_mgps <- bfa_mgps(sim$X)
}
