% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary_k}
\alias{summary_k}
\title{Summary for adaptive number of factors}
\usage{
summary_k(fit, real = NULL)
}
\arguments{
\item{fit}{Returned samples from any \verb{bfa_*()} with \code{adapt=TRUE}.}

\item{real}{Real number of factors to include in summary, \code{NULL} otherwise.}
}
\value{
Median number of factors in \code{fit$k}, also prints summary of the
posterior distribution of \code{k}.
}
\description{
Summary for adaptive number of factors
}
\examples{
set.seed(219)

# 0. Data simulation as in Bhattacharya & Dunson (2011)
sim <- sim_data_bfa(n = 200, p = 20, k = 3)

# 1. Adaptive number of factors, starting at floor(3*log(p)) = 8
fit_mgps <- bfa_mgps(sim$X)
k_fit <- summary_k(fit_mgps, real = sim$k_true)
}
