% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfa_cusp.R
\name{bfa_cusp}
\alias{bfa_cusp}
\title{Bayesian Factor Analysis and Covariance Estimation with a CUSP prior}
\usage{
bfa_cusp(
  X,
  iter_warmup = 1000,
  iter_sampling = 1000,
  adapt = TRUE,
  k_init = NULL,
  prior_params = list(a_sigma = NULL, b_sigma = NULL, alpha = NULL, a_theta = NULL,
    b_theta = NULL, theta_inf = NULL),
  verbose = TRUE
)
}
\arguments{
\item{X}{Data matrix}

\item{iter_warmup}{Number of warmup iterations to run per chain}

\item{iter_sampling}{Number of post-warmup iterations to run per chain}

\item{adapt}{(logical) Adapt number of factors across samples?}

\item{k_init}{If \code{adapt=FALSE}, the persistent number of factors. If
\code{adapt=TRUE}, the initial number of factors in the sampler. If \code{NULL}
the number of factors is selected automatically as \code{3*log(p)}, where \code{p}
is the number of columns of \code{X}}

\item{prior_params}{Parameters of the CUSP prior}

\item{verbose}{(logical) Show progress bar?}
}
\value{
List with samples for each parameter
}
\description{
Bayesian Factor Analysis and Covariance Estimation with a CUSP prior
}
\examples{
NULL
}
