% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfa_dl.R
\name{bfa_dl}
\alias{bfa_dl}
\title{Bayesian Factor Analysis and Covariance Estimation with a row-wise DL prior}
\usage{
bfa_dl(
  X,
  iter_warmup = 1000,
  iter_sampling = 1000L,
  k = NULL,
  prior_params = list(a_sigma = NULL, b_sigma = NULL, a = NULL),
  verbose = TRUE
)
}
\arguments{
\item{X}{Data matrix}

\item{iter_warmup}{Number of warmup iterations to run per chain.}

\item{iter_sampling}{Number of post-warmup iterations to run per chain.}

\item{k}{Number of factors (fixed)}

\item{prior_params}{Parameters of the DL prior.}

\item{verbose}{(logical) Show progress bar?}
}
\value{
List with samples for each parameter
}
\description{
Bayesian Factor Analysis and Covariance Estimation with a row-wise DL prior
}
\examples{
NULL
}
