% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfa_mgps.R
\name{bfa_mgps}
\alias{bfa_mgps}
\title{Bayesian Factor Analysis and Covariance Estimation with an MGPS prior}
\usage{
bfa_mgps(
  X,
  iter_warmup = 1000,
  iter_sampling = 1000,
  adapt = TRUE,
  k_init = NULL,
  prior_params = list(a_sigma = NULL, b_sigma = NULL, nu = NULL, a1 = NULL, a2 = NULL,
    alpha0 = NULL, alpha1 = NULL),
  eps = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{X}{Data matrix}

\item{iter_warmup}{Number of warmup iterations to run per chain.}

\item{iter_sampling}{Number of post-warmup iterations to run per chain.}

\item{adapt}{(logical) Adapt number of factors across samples?}

\item{k_init}{If \code{adapt = FALSE}, the persistent number of factors. If
\code{adapt = TRUE}, the initial number of factors in the sampler. If \code{NULL}
the number of factors is selected automatically as \code{3*log(p)} where \code{p}
is the number of columns of \code{X}.}

\item{prior_params}{Parameters of the MGPS prior.}

\item{eps}{If \code{adapt = TRUE}, tolerance for column sums of \code{Lambda} in the
adaptation step.}

\item{verbose}{(logical) Show progress bar?}
}
\value{
List with samples for each parameter
}
\description{
Bayesian Factor Analysis and Covariance Estimation with an MGPS prior
}
\examples{
set.seed(219)

# 0. Data simulation as in Bhattacharya & Dunson (2011)
sim <- sim_data_bfa(n = 200, p = 20, k = 3)

# 1. Adaptive number of factors, starting at floor(3*log(p)) = 8
fit_mgps <- bfa_mgps(sim$X)
k_fit <- summary_k(fit_mgps, real = sim$k_true) # returns median too
plots_Sigma_X <- plot_Sigma_X(fit_mgps, real = sim$Sigma_X_true)
plots_Sigma_X$main
plots_Sigma_X$residual
plots_Sigma_X$scatter

# 2. Fixed number of factors -> MatchAlign
fit_mgps_fixed <- bfa_mgps(sim$X, adapt = FALSE, k_init = k_fit)

# TODO: add MatchAlign
}
