% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blfr_dl.R
\name{blfr_dl}
\alias{blfr_dl}
\title{Bayesian Latent Factor Regression with a Dirichlet-Laplace prior}
\usage{
blfr_dl(
  y,
  X,
  iter_warmup = 1000,
  iter_sampling = 1000,
  k = NULL,
  induced = TRUE,
  interactions = FALSE,
  prior_params = list(nu_y = NULL, beta_var = NULL, Omega_var = NULL, a_sigma = NULL,
    b_sigma = NULL, a = NULL),
  mala_eps = NULL,
  adapt_mala_eps = TRUE,
  window_mala = floor(iter_warmup/10),
  verbose = TRUE
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Predictors matrix}

\item{iter_warmup}{Number of warmup iterations to run per chain}

\item{iter_sampling}{Number of post-warmup iterations to run per chain}

\item{k}{Number of factors}

\item{induced}{(logical) return induced effects of original predictors?}

\item{interactions}{(logical) include interactions between latent factors?}

\item{prior_params}{Parameters of the Dirichlet-Laplace prior}

\item{mala_eps}{If \code{interactions=TRUE}, step-size of the MALA step for
latent factors. If \code{NULL}, uses \code{k^(-1/3)} as default.}

\item{adapt_mala_eps}{(logical) Should \code{mala_eps} be adaptive?}

\item{window_mala}{If \code{adapt_mala_eps = TRUE}, number of iterations during
warmup that are considered for the adaptation of \code{mala_eps}}

\item{verbose}{(logical) Show progress bar?}
}
\value{
List with samples for each parameter
}
\description{
Bayesian Latent Factor Regression with a Dirichlet-Laplace prior
}
\examples{
NULL
}
