% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blfr_mgps.R
\name{blfr_mgps}
\alias{blfr_mgps}
\title{Bayesian Latent Factor Regression with a MGPS prior}
\usage{
blfr_mgps(
  y,
  X,
  iter_warmup = 1000,
  iter_sampling = 1000,
  k = NULL,
  induced = TRUE,
  interactions = FALSE,
  prior_params = list(nu_y = NULL, beta_var = NULL, Omega_var = NULL, a_sigma = NULL,
    b_sigma = NULL, nu = NULL, a1 = NULL, a2 = NULL),
  mala_eps = NULL,
  adapt_mala_eps = TRUE,
  window_mala = floor(iter_warmup/10),
  verbose = TRUE
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Predictors matrix}

\item{iter_warmup}{Number of warmup iterations to run per chain.}

\item{iter_sampling}{Number of post-warmup iterations to run per chain.}

\item{k}{Number of factors}

\item{induced}{(logical) return induced effects of original predictors?}

\item{interactions}{(logical) include interactions between latent factors?}

\item{prior_params}{Parameters of the MGPS prior}

\item{mala_eps}{If \code{interactions=TRUE}, step-size of the MALA step for
latent factors. If \code{NULL}, uses \code{k^(-1/3)} as default.}

\item{adapt_mala_eps}{(logical) Should \code{mala_eps} be adaptive?}

\item{window_mala}{If \code{adapt_mala_eps = TRUE}, number of iterations during
warmup that are considered for the adaptation of \code{mala_eps}}

\item{verbose}{(logical) Show progress bar?}
}
\value{
List with samples for each parameter
}
\description{
Bayesian Latent Factor Regression with a MGPS prior
}
\examples{
set.seed(219)

# 1. Example without interactions
sim <- sim_data_blfr(n = 200, n_test = 100, p = 20, k = 3)
fit_mgps <- blfr_mgps(sim$y, sim$X)

# 1.1 Induced effect summaries
Sigma_X_plots <- plot_Sigma_X(fit_mgps, real = sim$Sigma_X_true)
Sigma_X_plots$scatter
main_effects_plots <- plot_effects_X(
    fit_mgps,
    real_beta = sim$beta_X
)
main_effects_plots$main
main_effects_plots$scatter

# 1.2 Match Align
aligned <- match_align(fit_mgps)
plot_match_align(aligned, "Lambda", "mean")

# 1.3 Estimates on test set
preds <- predict_blfr(fit_mgps, sim$X_test)
mean((preds - sim$y_test)^2)
plot(1:length(preds), preds - sim$y_test)

# 2. Example with interactions
#sim <- sim_data_blfr(200, 100, 20, k = 3, interactions = TRUE)
#fit_mgps <- blfr_mgps(sim$y, sim$X, interactions = TRUE)

# 2.1 Induced effect summaries
# Sigma_X_plots <- plot_Sigma_X(fit_mgps, real = sim$Sigma_X_true)
# Sigma_X_plots$scatter
# main_effects_plots <- plot_effects_X(
#   fit_mgps, real_beta = sim$beta_X, real_intercept = sim$intercept_X
# )
# main_effects_plots$main
# interaction_effects_plots <- plot_Omega_X(fit_mgps, real = sim$Omega_X)

# 2.2 Match Align
# aligned <- match_align(fit_mgps)
# plot_match_align(aligned, "Lambda", "mean")

}
