% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blfr_mgps.R
\name{blfr_mgps}
\alias{blfr_mgps}
\title{Bayesian Latent Factor Regression with an MGPS prior}
\usage{
blfr_mgps(
  y,
  X,
  iter_warmup = 1000,
  iter_sampling = 1000,
  k = NULL,
  induced = TRUE,
  interactions = FALSE,
  prior_params = list(nu_y = NULL, beta_var = NULL, Omega_var = NULL, a_sigma = NULL,
    b_sigma = NULL, nu = NULL, a1 = NULL, a2 = NULL),
  mala_eps = 10,
  verbose = TRUE
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Predictors matrix}

\item{iter_warmup}{Number of warmup iterations to run per chain.}

\item{iter_sampling}{Number of post-warmup iterations to run per chain.}

\item{k}{Number of factors}

\item{induced}{(logical) return induced effects of original predictors?}

\item{interactions}{(logical) include interactions between latent factors?}

\item{prior_params}{Parameters of the MGPS prior}

\item{mala_eps}{If \code{interactions=TRUE}, step-size of the MALA step for
latent factors}

\item{verbose}{(logical) Show progress bar?}
}
\value{
List with samples for each parameter
}
\description{
Bayesian Latent Factor Regression with an MGPS prior
}
\examples{
NULL
}
