% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_effects_X}
\alias{plot_effects_X}
\title{Summary Plots of Induced Main Effects}
\usage{
plot_effects_X(fit_blfr, real_beta = NULL, real_intercept = NULL)
}
\arguments{
\item{fit_blfr}{Returned samples from any \verb{blfr_*()} function}

\item{real_beta}{Real vector of main effects, if known}

\item{real_intercept}{Real intercept, if known}
}
\value{
Named list of plots
}
\description{
Summary Plots of Induced Main Effects
}
\examples{
NULL
}
