% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_align_algorithm.R
\name{plot_match_align}
\alias{plot_match_align}
\title{Plots of results of MatchAlign}
\usage{
plot_match_align(
  match_align_out,
  type = c("Lambda", "Eta"),
  summary = c("mean", "median"),
  color = "green",
  title = NULL,
  X_names = NULL
)
}
\arguments{
\item{match_align_out}{Output of \code{match_align()}.}

\item{type}{Which samples to plot? One of \code{c("Lambda", "Eta")}.}

\item{summary}{Which summary metric to use? One of \code{c("mean", "median")}.}

\item{color}{Color scheme. One of \code{c("green", "red", "wes")}}

\item{title}{Optional plot title}

\item{X_names}{Optional character vector of covariate names, of length equal
to the number of variables \code{p}. When supplied (and the plotted matrix has
\code{p} rows, i.e. \code{type = "Lambda"}), these are used to label the y axis.}
}
\value{
ggplot object
}
\description{
Plots the summary matrix after performing a rotation of the factors and
loadings to induce joint identifiability.
}
\examples{
set.seed(219)

# 0. Data simulation as in Bhattacharya & Dunson (2011)
sim <- sim_data_bfa(n = 200, p = 20, k = 3)

# 1. Adaptive number of factors, starting at floor(3*log(p)) = 8
fit_cusp <- bfa_cusp(sim$X)
k_fit <- summary_k(fit_cusp, real = sim$k_true) # returns median too

# 2. Fixed number of factors -> MatchAlign
fit_cusp_fixed <- bfa_mgps(sim$X, adapt = FALSE, k_init = k_fit)
aligned <- match_align(fit_cusp_fixed)
plot_match_align(aligned, "Lambda", "mean")
plot_match_align(aligned, "Eta", "mean")
}
