% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_Sigma_X}
\alias{plot_Sigma_X}
\title{Summary Plots of Posterior Covariance}
\usage{
plot_Sigma_X(fit, real = NULL, X_names = NULL)
}
\arguments{
\item{fit}{Returned samples from any \verb{bfa_*()} or \verb{blfr_*()} function.}

\item{real}{Real Sigma_X if known, \code{NULL} otherwise}

\item{X_names}{Optional character vector of covariate names, of length equal
to the number of variables in \code{X}. If supplied, these are used to label the
x and y axes of the covariance heatmaps.}
}
\value{
Named list of plots. If \code{real=NULL}, \code{plots$main} is the plot of the
mean posterior covariance of X. Otherwise, three plots:
\itemize{
\item \code{main}: compares mean posterior covariance with real covariance
\item \code{residual}: plot of residuals between mean posterior and real covariance
\item \code{scatter}: scatterplot of mean posterior elements vs real elements.
}
}
\description{
Summary Plots of Posterior Covariance
}
\examples{
set.seed(219)

# 0. Data simulation as in Bhattacharya & Dunson (2011)
sim <- sim_data_bfa(n = 200, p = 20, k = 3)

# 1. Adaptive number of factors, starting at floor(3*log(p)) = 8
fit_mgps <- bfa_mgps(sim$X)
plots_Sigma_X <- plot_Sigma_X(fit_mgps, real = sim$Sigma_X_true)
plots_Sigma_X$main
plots_Sigma_X$residual
plots_Sigma_X$scatter
}
