% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data_bfa}
\alias{sim_data_bfa}
\title{Generate Data for Factor Analysis Examples}
\usage{
sim_data_bfa(n, p, k)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of covariates}

\item{k}{Number of latent factors}
}
\value{
Named list with the following elements:
\itemize{
\item \code{X}: \eqn{n \times p} matrix of observations
\item \code{Lambda}: Simulated factor loadings
\item \code{sigma2}: Vector of intrinsic variances for each covariate
\item \code{Sigma_X_true}: True covariance matrix of \code{X}, i.e.
\eqn{\Lambda^T\lambda + \text{diag}(\sigma^2)}
\item \code{k_true}: True number of latent factors
}
}
\description{
Data generation as in Bhattacharya & Dunson (2011)
}
\details{
Data is generated as in Bhattacharya & Dunson (2011): number of non-zero
columns of the matrix of loadings is linearly decreasing from 2k down to
k + 1 and selected at random. Then non-zero elements are sampled
from a N(0, 9) distribution. Note that this is a simulation from any
"true" model.
}
\examples{
set.seed(219)

# 0. Data simulation as in Bhattacharya & Dunson (2011)
sim <- sim_data_bfa(n = 200, p = 20, k = 3)

# 1. Adaptive number of factors, starting at floor(3*log(p)) = 8
fit_mgps <- bfa_mgps(sim$X)
}
